requirements.txt
distribute==0.6.27
nose==1.1.2
pyhamcrest==1.6
should-dsl==2.0a4
